team_service/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
pymongo[srv]==3.12.1
djongo==1.3.6
pytest-django==4.5.2
django-environ==0.9.0
dj-rest-auth==2.2.5
django-allauth==0.51.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
google-auth==2.15.0
sqlparse==0.2.4
ruamel.yaml.clib==0.2.2
drf_yasg==1.21.4
django-redis==5.2.0